feat: Add default category and point earning rule to seed_database.py
This commit adds a default category and point earning rule to the `seed_database.py` file. The default category is added with an ID of 0 and the name "Default". The default point earning rule is added with the default category, a points per dollar value of 1, and a start date of January 1, 1900 and an end date of December 31, 2099.